+++ src/app/[locale]/layout.config.tsx
@@ -7,26 +7,29 @@ import { BaseLayoutProps } from 'fumadocs-ui/layouts/shared';
 import { getTranslations } from 'next-intl/server';
 import { CreditPopover } from '@/components/credit-popover';
 import { ExtendedLinkItem, HomeTitle } from '@windrun-huaiin/third-ui/fuma/base';
+import { getOptionalAuth } from '@windrun-huaiin/third-ui/clerk/patch/optional-auth';
+import { getAsNeededLocalizedUrl } from '@windrun-huaiin/lib';
 
 // home page normal menu
 export async function homeNavLinks(locale: string): Promise<ExtendedLinkItem[]> {
   const t1 = await getTranslations({ locale: locale, namespace: 'linkPreview' });
+  const { userId } = await getOptionalAuth();
   return [
     {
       icon: <icons.BugOff />,
       text: t1('blog'),
-      url: `/${locale}/blog`,
+      url: getAsNeededLocalizedUrl(locale, '/blog'),
     },
     {
       icon: <icons.BTC />,
       text: t1('pricing'),
-      url: `/${locale}/pricing`,
+      url: getAsNeededLocalizedUrl(locale, '/pricing'),
     },
     {
       type: 'custom',
       secondary: true,
       mobilePinned: true,
-      children: <CreditPopover locale={locale} />,
+      children: userId ? <CreditPopover locale={locale} /> : null,
     },
     {
       type: 'custom',
@@ -49,7 +52,7 @@ export async function baseOptions(locale: string): Promise<BaseLayoutProps> {
   const t = await getTranslations({ locale: locale, namespace: 'home' });
   return {
     nav: {
-      url: `/${locale}`,
+      url: getAsNeededLocalizedUrl(locale, '/'),
       title: (
         <>
           <SiteIcon />