+++ src/app/[locale]/layout.tsx
@@ -8,6 +8,7 @@ import { RootProvider } from "fumadocs-ui/provider/next";
 import { ClerkProviderClient } from '@windrun-huaiin/third-ui/clerk';
 import { NextIntlClientProvider } from 'next-intl';
 import { getMessages, getTranslations, setRequestLocale } from 'next-intl/server';
+import { getAsNeededLocalizedUrl } from '@windrun-huaiin/lib';
 import React from 'react';
 import './globals.css';
 
@@ -27,9 +28,10 @@ export async function generateMetadata({
     keywords: t('keywords'),
     metadataBase: new URL(appConfig.baseUrl),
     alternates: {
-      canonical: `${appConfig.baseUrl}/${locale}`,
+      canonical: `${appConfig.baseUrl}${getAsNeededLocalizedUrl(locale, '/')}`,
       languages: {
-        "en": `${appConfig.baseUrl}/en`,
+        "en": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('en', '/')}`,
+        "zh": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('zh', '/')}`,
       }
     },
     icons: [